+++ MindKey.Server/Models/IUploadRepository.cs
@@ -6,9 +6,9 @@ namespace MindKey.Server.Models
     public interface IUploadRepository
     {
         PagedResult<Upload> GetUploads(string? name, int page);
-        Task<Upload?> GetUpload(int Id);
+        Task<Upload?> GetUpload(long Id);
         Task<Upload> AddUpload(Upload upload);
         Task<Upload?> UpdateUpload(Upload upload);
-        Task<Upload?> DeleteUpload(int id);
+        Task<Upload?> DeleteUpload(long id);
     }
 }
